Added 05.05

+++ src/main/java/pl/coderslab/controller/InvestityController.java
@@ -40,20 +40,20 @@ public class InvestityController {
 
     @GetMapping(path = "admin/investity/add")
     @Secured("ROLE_ADMIN")
-    String addNewInvestity(Model model){
+    String addNewInvestity(Model model) {
         Investity investity = new Investity();
-        model.addAttribute("investity",investity);
+        model.addAttribute("investity", investity);
         return "admin/investity/add";
     }
 
     @PostMapping(path = "admin/investity/add")
     @Secured("ROLE_ADMIN")
-    String addNewInvestity(@Valid Investity investity, BindingResult result){
-        if(investityService.isInvestityNameTaken(investity)){
+    String addNewInvestity(@Valid Investity investity, BindingResult result) {
+        if (investityService.isInvestityNameTaken(investity)) {
             result.addError(new FieldError("investity", "investityName"
                     , "Nazwa inwestycji jest zajęta!"));
         }
-        if(result.hasErrors()){
+        if (result.hasErrors()) {
             return "admin/investity/add";
         }
         investityService.addNewInvestity(investity);
@@ -62,33 +62,33 @@ public class InvestityController {
 
     @GetMapping(path = "admin/investity/all")
     @Secured("ROLE_ADMIN")
-    String showAllInvestities(Model model){
+    String showAllInvestities(Model model) {
         List<Investity> investities = investityService.findAll();
-        Map<Long,Double> allCosts = investityService.getAllCosts();
-        Map<Long,Double> allMargins = investityService.getAllMargins();
-        model.addAttribute("investities",investities);
+        Map<Long, Double> allCosts = investityService.getAllCosts();
+        Map<Long, Double> allMargins = investityService.getAllMargins();
+        model.addAttribute("investities", investities);
         model.addAttribute("allCosts", allCosts);
-        model.addAttribute("allMargins",allMargins);
+        model.addAttribute("allMargins", allMargins);
         return "admin/investity/all";
     }
 
     @GetMapping(path = "admin/investity/edit")
     @Secured("ROLE_ADMIN")
-    String editInvestity(Model model, @RequestParam long id){
+    String editInvestity(Model model, @RequestParam long id) {
         Optional<Investity> investity = investityService.findById(id);
         model.addAttribute("investity"
-                ,investity.orElseThrow(()->new EntityNotFoundException("Could not found investity " + id)));
+                , investity.orElseThrow(() -> new EntityNotFoundException("Could not found investity " + id)));
         return "admin/investity/edit";
     }
 
     @PostMapping(path = "admin/investity/edit")
     @Secured("ROLE_ADMIN")
-    String editInvestity(@Valid Investity investity, BindingResult result){
-        if(investityService.isInvestityNameTaken(investity)){
+    String editInvestity(@Valid Investity investity, BindingResult result) {
+        if (investityService.isInvestityNameTaken(investity)) {
             result.addError(new FieldError("investity", "investityName"
                     , "Nazwa inwestycji jest zajęta!"));
         }
-        if(result.hasErrors()){
+        if (result.hasErrors()) {
             return "admin/investity/add";
         }
         investityService.editInvestity(investity);
@@ -97,16 +97,19 @@ public class InvestityController {
 
     @GetMapping(path = "admin/investity/delete")
     @Secured("ROLE_ADMIN")
-    String deleteInvestity(@RequestParam long id){
+    String deleteInvestity(@RequestParam long id) {
+        if (investityCostsService.findAllByInvestityId(id).size() > 0 || workingTimeService.findAllByInvestityId(id).size() > 0) {
+            return "redirect:/admin/error?id=1";
+        }
         investityService.deleteInvestityById(id);
         return "redirect:/admin/investity/all";
     }
 
     @GetMapping(path = "admin/investity/info")
     @Secured("ROLE_ADMIN")
-    String showInfo(@RequestParam long id, Model model){
+    String showInfo(@RequestParam long id, Model model) {
         Investity investity = investityService.findById(id)
-                .orElseThrow(()->new EntityNotFoundException("Could not found investity " + id));
+                .orElseThrow(() -> new EntityNotFoundException("Could not found investity " + id));
         List<InvestityCosts> investityCosts = investityCostsService.findAllByInvestityId(id);
         int sumOfSuppliersCosts = investityCostsService.sumOfAllInvestityCosts(id);
         int sumOfUserCosts = workingTimeService.getAllUserCosts(id);
@@ -119,82 +122,86 @@ public class InvestityController {
                 = workingTimeService.getAllHoursByUserInInvestityWithOtherMultiplier(id);
         Map<Long, Double> getAllCostsByUserInInvestity = workingTimeService.getAllCostsByUserInInvestity(id);
 
-        model.addAttribute("investity",investity);
-        model.addAttribute("investityCosts",investityCosts);
-        model.addAttribute("sumOfSuppliersCosts",sumOfSuppliersCosts);
-        model.addAttribute("margin",margin);
+        model.addAttribute("investity", investity);
+        model.addAttribute("investityCosts", investityCosts);
+        model.addAttribute("sumOfSuppliersCosts", sumOfSuppliersCosts);
+        model.addAttribute("margin", margin);
 
-        model.addAttribute("sumOfUserCosts",sumOfUserCosts);
-        model.addAttribute("allUsers",allUsers);
-        model.addAttribute("allHoursByUserInInvestityMultiplierOne",allHoursByUserInInvestityMultiplierOne);
-        model.addAttribute("allHoursByUserInInvestityMultiplierOther",allHoursByUserInInvestityMultiplierOther);
-        model.addAttribute("getAllCostsByUserInInvestity",getAllCostsByUserInInvestity);
+        model.addAttribute("sumOfUserCosts", sumOfUserCosts);
+        model.addAttribute("allUsers", allUsers);
+        model.addAttribute("allHoursByUserInInvestityMultiplierOne", allHoursByUserInInvestityMultiplierOne);
+        model.addAttribute("allHoursByUserInInvestityMultiplierOther", allHoursByUserInInvestityMultiplierOther);
+        model.addAttribute("getAllCostsByUserInInvestity", getAllCostsByUserInInvestity);
         return "admin/investity/info";
     }
 
     @GetMapping(path = "admin/investity/info/editcost")
     @Secured("ROLE_ADMIN")
-    String editInvestityCosts(@RequestParam long id, Model model){
+    String editInvestityCosts(@RequestParam long id, Model model) {
         Optional<InvestityCosts> investityCosts = investityCostsService.findCosts(id);
         model.addAttribute("investityCosts", investityCosts
-                .orElseThrow(()-> new EntityNotFoundException("Could not found investity costs " + id)));
+                .orElseThrow(() -> new EntityNotFoundException("Could not found investity costs " + id)));
         return "admin/investity/info/editcost";
     }
+
     @PostMapping(path = "admin/investity/info/editcost")
     @Secured("ROLE_ADMIN")
-    String editInvestityCosts(@Valid InvestityCosts investityCosts, BindingResult result){
-        if(result.hasErrors()){
+    String editInvestityCosts(@Valid InvestityCosts investityCosts, BindingResult result) {
+        if (result.hasErrors()) {
             return "admin/investity/info/editcost";
         }
         investityCostsService.save(investityCosts);
-        return "redirect:/admin/investity/info?id="+investityCosts.getInvestity().getId();
+        return "redirect:/admin/investity/info?id=" + investityCosts.getInvestity().getId();
     }
 
     @GetMapping(path = "admin/investity/info/deletecost")
     @Secured("ROLE_ADMIN")
-    String deleteInvestityCosts(@RequestParam long id){
+    String deleteInvestityCosts(@RequestParam long id) {
         long redirectId = investityCostsService.findCosts(id).orElseThrow(
-                ()->new EntityNotFoundException("Could not found investity costs " + id)).getInvestity().getId();
+                () -> new EntityNotFoundException("Could not found investity costs " + id)).getInvestity().getId();
         investityCostsService.deleteById(id);
-        return "redirect:/admin/investity/info?id="+redirectId;
+        return "redirect:/admin/investity/info?id=" + redirectId;
     }
+
     @GetMapping(path = "admin/investity/info/addcost")
     @Secured("ROLE_ADMIN")
-    String addInvestityCosts(Model model, @RequestParam long investityId){
+    String addInvestityCosts(Model model, @RequestParam long investityId) {
         InvestityCosts investityCosts = new InvestityCosts();
         investityCosts.setInvestity(investityService.findById(investityId)
-                .orElseThrow(()->new EntityNotFoundException("Could not found investity " + investityId)));
-        model.addAttribute("investityCosts",investityCosts);
+                .orElseThrow(() -> new EntityNotFoundException("Could not found investity " + investityId)));
+        model.addAttribute("investityCosts", investityCosts);
         return "admin/investity/info/addcost";
     }
+
     @PostMapping(path = "admin/investity/info/addcost")
     @Secured("ROLE_ADMIN")
-    String addInvestityCosts(@Valid InvestityCosts investityCosts, BindingResult result){
-        if(result.hasErrors()){
+    String addInvestityCosts(@Valid InvestityCosts investityCosts, BindingResult result) {
+        if (result.hasErrors()) {
             return "admin/investity/info/addcost";
         }
         investityCostsService.save(investityCosts);
-        return "redirect:/admin/investity/info?id="+investityCosts.getInvestity().getId();
+        return "redirect:/admin/investity/info?id=" + investityCosts.getInvestity().getId();
     }
 
     @GetMapping(path = "admin/investity/info/showuserinfo")
     @Secured("ROLE_ADMIN")
-    String showUserInfoInInvestity(@RequestParam long userid, @RequestParam long investityid, Model model){
-        List<WorkingTime> workingTime = workingTimeService.findByInvestity_IdAndUser_Id(investityid,userid);
-        model.addAttribute("workingTime",workingTime);
-        model.addAttribute("user",userService.findByUserId(userid)
-                .orElseThrow(()->new EntityNotFoundException("Could not found user " + userid)));
-        model.addAttribute("investity",investityService.findById(investityid)
-                .orElseThrow(()->new EntityNotFoundException("Could not found investity " + investityid)));
+    String showUserInfoInInvestity(@RequestParam long userid, @RequestParam long investityid, Model model) {
+        List<WorkingTime> workingTime = workingTimeService.findByInvestity_IdAndUser_Id(investityid, userid);
+        model.addAttribute("workingTime", workingTime);
+        model.addAttribute("user", userService.findByUserId(userid)
+                .orElseThrow(() -> new EntityNotFoundException("Could not found user " + userid)));
+        model.addAttribute("investity", investityService.findById(investityid)
+                .orElseThrow(() -> new EntityNotFoundException("Could not found investity " + investityid)));
         return "admin/investity/info/showuserinfo";
     }
 
     @ModelAttribute("suppliers")
-    Collection<Supplier> findAllSuppliers(){
+    Collection<Supplier> findAllSuppliers() {
         return supplierService.findAll();
     }
+
     @ModelAttribute("dayOfWeekMap")
-    Map<DayOfWeek, String> changeDayOfWeekNameToPolish(){
+    Map<DayOfWeek, String> changeDayOfWeekNameToPolish() {
         Map<DayOfWeek, String> map = new HashMap<>();
         map.put(DayOfWeek.MONDAY, "Poniedziałek");
         map.put(DayOfWeek.TUESDAY, "Wtorek");
@@ -207,20 +214,20 @@ public class InvestityController {
     }
 
     @ModelAttribute("monthMap")
-    Map<Month, String> changeMonthNameToPolish(){
+    Map<Month, String> changeMonthNameToPolish() {
         Map<Month, String> map = new HashMap<>();
         map.put(Month.JANUARY, "Styczeń");
         map.put(Month.FEBRUARY, "Luty");
-        map.put(Month.MARCH , "Marzec");
-        map.put(Month.APRIL , "Kwiecień");
-        map.put(Month.MAY , "Maj");
-        map.put(Month.JUNE , "Czerwiec");
-        map.put(Month.JULY , "Lipiec");
-        map.put(Month.AUGUST , "Sierpień");
-        map.put(Month.SEPTEMBER , "Wrzesień");
-        map.put(Month.OCTOBER , "Październik");
-        map.put(Month.NOVEMBER , "Listopad");
-        map.put(Month.DECEMBER , "Grudzień");
+        map.put(Month.MARCH, "Marzec");
+        map.put(Month.APRIL, "Kwiecień");
+        map.put(Month.MAY, "Maj");
+        map.put(Month.JUNE, "Czerwiec");
+        map.put(Month.JULY, "Lipiec");
+        map.put(Month.AUGUST, "Sierpień");
+        map.put(Month.SEPTEMBER, "Wrzesień");
+        map.put(Month.OCTOBER, "Październik");
+        map.put(Month.NOVEMBER, "Listopad");
+        map.put(Month.DECEMBER, "Grudzień");
         return map;
     }
 }
